Added variable for character class
I added a public int for character class type that the script now chekcs for when the player uses attack inputs. CharacterClassType with a value of 1, for example, is the character we already have in the game (sword guy with basic light/heavy slashes). When the player hits an attack input, the script checks the CharacterClassType to make sure it's activating the right moveset.